refactor: flatten Tutorial framework package
