+++ tools/tutorial/gox.mod
@@ -1,3 +1,3 @@
 xgo 1.7.5
 
-project _course.gox Course github.com/goplus/builder/tools/tutorial/framework
+project _course.gox Course github.com/goplus/builder/tools/tutorial
